+++ src/rest_framework_tricks/utils.py
@@ -10,7 +10,7 @@ __license__ = "GPL-2.0-only OR LGPL-2.1-or-later"
 __all__ = ("DictProxy",)
 
 
-class DictProxy(object):
+class DictProxy:
     """Dictionary proxy.
 
     Example: